ai-engine/requirements.txt
# AI/ML Libraries
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pytesseract==0.3.10

# Image Processing
opencv-python-headless==4.8.1.78
imutils==0.5.4

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Redis
redis==5.0.1

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0

# Development
black==23.11.0
isort==5.12.0
